HW 6: схемы в отдельном подуле Schemas
