+++ part6many_to_many/lib/part6many_to_many/todo_list_item.ex
@@ -1,26 +1,10 @@
-defmodule Part6manyToMany.TodoListItems do
-  use Ecto.Schema
-  import Ecto.Changeset
-
-  @primary_key false
-  schema "todo_list_items" do
-    belongs_to :todo_list, Part6manyToMany.TodoList
-    belongs_to :todo_item, Part6manyToMany.TodoItem
-    timestamps()
-  end
-
+defmodule Part6manyToMany.TodoListItem do
   import Ecto.Query, warn: false
   alias Part6manyToMany.Repo
-  alias Part6manyToMany.TodoListItem
-  alias Part6manyToMany.TodoList
-  alias Part6manyToMany.TodoItem
 
-  @doc false
-  def changeset(attrs), do: cast(attrs)
-
-  def cast(list, item), do: {:ok, list, item}
-
-  def cast(attrs), do: {:ok, attrs}
+  alias Part6manyToMany.Schemas.TodoListItem
+  alias Part6manyToMany.Schemas.TodoList
+  alias Part6manyToMany.Schemas.TodoItem
 
   @doc """
   Returns the list of relations.
