+++ part6many_to_many/test/support/fixtures/todo_list_fixtures.ex
@@ -1,4 +1,6 @@
 defmodule Part6manyToMany.TodoListFixtures do
+  alias Part6manyToMany.TodoList
+
   @moduledoc """
   This module defines test helpers for creating
   entities via the `Part6manyToMany.TodoList` context.
@@ -8,7 +10,7 @@ defmodule Part6manyToMany.TodoListFixtures do
   Generate a todo_list.
   """
   def todo_list_fixture(title = %{title: entry_title} \\ %{title: "Default Title Fixture"}) do
-    {:ok, todo_list} = Part6manyToMany.TodoList.create_todo_list(title)
+    {:ok, todo_list} = TodoList.create_todo_list(title)
 
     todo_list
   end
